fix: rename javascript runtime to script runtime
