+++ packages/core/src/workspace/index.ts
@@ -1,3 +1,4 @@
 export * from './modules/index.js';
 export * from './YasumuWorkspace.js';
 export * from './YasumuWorkspaceMetadata.js';
+export * from './runtime/index.js';
