+++ packages/core/src/workspace/runtime/index.ts
@@ -1,3 +1,3 @@
-export * from './BaseJavaScriptRuntime.js';
+export * from './BaseScriptRuntime.js';
 export * from './TanxiumRuntime.js';
 export * from './WebWorkerRuntime.js';
